refactor(cas): 权限管理重构

+++ dunwu-admin/dunwu-admin-modules/dunwu-admin-module-security/src/main/java/io/github/dunwu/module/cas/service/impl/MenuServiceImpl.java
@@ -1,19 +1,21 @@
 package io.github.dunwu.module.cas.service.impl;
 
 import cn.hutool.core.bean.BeanUtil;
+import cn.hutool.core.bean.copier.CopyOptions;
 import cn.hutool.core.collection.CollectionUtil;
 import cn.hutool.core.util.ObjectUtil;
 import cn.hutool.core.util.StrUtil;
-import com.baomidou.mybatisplus.core.toolkit.Wrappers;
+import io.github.dunwu.module.cas.constant.enums.PermissionType;
 import io.github.dunwu.module.cas.dao.MenuDao;
-import io.github.dunwu.module.cas.dao.RoleMenuMapDao;
+import io.github.dunwu.module.cas.dao.PermissionDao;
 import io.github.dunwu.module.cas.entity.Menu;
-import io.github.dunwu.module.cas.entity.RoleMenuMap;
+import io.github.dunwu.module.cas.entity.Permission;
 import io.github.dunwu.module.cas.entity.dto.MenuDto;
 import io.github.dunwu.module.cas.entity.dto.RoleDto;
 import io.github.dunwu.module.cas.entity.query.MenuQuery;
-import io.github.dunwu.module.cas.entity.vo.MenuVo;
+import io.github.dunwu.module.cas.entity.vo.VueElementMenuVo;
 import io.github.dunwu.module.cas.service.MenuService;
+import io.github.dunwu.module.cas.service.PermissionService;
 import io.github.dunwu.module.cas.service.RoleService;
 import io.github.dunwu.tool.data.mybatis.ServiceImpl;
 import io.github.dunwu.tool.util.tree.Node;
@@ -25,6 +27,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.io.Serializable;
 import java.util.*;
@@ -35,55 +38,121 @@ import javax.servlet.http.HttpServletResponse;
  * 菜单表 Service 类
  *
  * @author <a href="mailto:forbreak@163.com">Zhang Peng</a>
- * @since 2021-10-12
+ * @since 2021-10-16
  */
 @Service
 @RequiredArgsConstructor
 public class MenuServiceImpl extends ServiceImpl implements MenuService {
 
     private final MenuDao menuDao;
-    private final RoleMenuMapDao roleMenuDao;
+    private final PermissionDao permissionDao;
     private final RoleService roleService;
+    private final PermissionService permissionService;
     private final SecurityService securityService;
+    private static final CopyOptions options;
+
+    static {
+        options = new CopyOptions();
+        options.setIgnoreProperties(Menu.ID, Menu.PID);
+    }
 
     @Override
+    @Transactional(rollbackFor = Exception.class)
     public boolean insert(Menu entity) {
-        return menuDao.insert(entity);
+        boolean isOk = menuDao.insert(entity);
+        if (!isOk) {
+            return false;
+        }
+
+        Permission permission = BeanUtil.toBean(entity, Permission.class, options);
+        permission.setResourceId(entity.getId());
+        permission.setType(PermissionType.MENU.getCode());
+        return permissionDao.insert(permission);
     }
 
     @Override
+    @Transactional(rollbackFor = Exception.class)
     public boolean insertBatch(Collection<Menu> list) {
-        return menuDao.insertBatch(list);
+        if (CollectionUtil.isEmpty(list)) {
+            return false;
+        }
+
+        list.forEach(this::insert);
+        return true;
     }
 
     @Override
+    @Transactional(rollbackFor = Exception.class)
     public boolean updateById(Menu entity) {
-        return menuDao.updateById(entity);
+        boolean isOk = menuDao.updateById(entity);
+        if (!isOk) {
+            return false;
+        }
+
+        Permission permission = BeanUtil.toBean(entity, Permission.class, options);
+        permission.setResourceId(entity.getId());
+        permission.setType(PermissionType.MENU.getCode());
+        return permissionDao.saveByResourceId(permission);
     }
 
     @Override
+    @Transactional(rollbackFor = Exception.class)
     public boolean updateBatchById(Collection<Menu> list) {
-        return menuDao.updateBatchById(list);
+        if (CollectionUtil.isEmpty(list)) {
+            return false;
+        }
+
+        list.forEach(this::updateById);
+        return true;
     }
 
     @Override
+    @Transactional(rollbackFor = Exception.class)
     public boolean save(Menu entity) {
-        return menuDao.save(entity);
+        boolean isOk = menuDao.save(entity);
+        if (!isOk) {
+            return false;
+        }
+
+        Permission permission = BeanUtil.toBean(entity, Permission.class, options);
+        permission.setResourceId(entity.getId());
+        permission.setType(PermissionType.MENU.getCode());
+        return permissionDao.saveByResourceId(permission);
     }
 
     @Override
+    @Transactional(rollbackFor = Exception.class)
     public boolean saveBatch(Collection<Menu> list) {
-        return menuDao.saveBatch(list);
+        if (CollectionUtil.isEmpty(list)) {
+            return false;
+        }
+
+        list.forEach(this::save);
+        return true;
     }
 
     @Override
     public boolean deleteById(Serializable id) {
-        return menuDao.deleteById(id);
+        boolean isOk = menuDao.deleteById(id);
+        if (!isOk) {
+            return false;
+        }
+
+        return permissionDao.deleteByResourceId(id);
     }
 
     @Override
     public boolean deleteBatchByIds(Collection<? extends Serializable> ids) {
-        return menuDao.deleteBatchByIds(ids);
+        if (CollectionUtil.isEmpty(ids)) {
+            return false;
+        }
+
+        boolean isOk = menuDao.deleteBatchByIds(ids);
+        if (!isOk) {
+            return false;
+        }
+
+        return permissionDao.deleteBatchByResourceIds(ids);
     }
 
     @Override
@@ -145,10 +214,11 @@ public class MenuServiceImpl extends ServiceImpl implements MenuService {
             Map<String, Object> map = new LinkedHashMap<>();
             map.put("ID", item.getId());
             map.put("上级菜单ID", item.getPid());
-            map.put("子菜单数目", item.getSubCount());
-            map.put("菜单类型", item.getType());
-            map.put("菜单标题", item.getTitle());
-            map.put("组件名称", item.getName());
+            map.put("子菜单数目", item.getChildrenNum());
+            map.put("编码", item.getCode());
+            map.put("名称", item.getName());
+            map.put("菜单类型", item.getMenuType());
+            map.put("权限表达式", item.getExpression());
             map.put("组件", item.getComponent());
             map.put("排序", item.getSequence());
             map.put("图标", item.getIcon());
@@ -156,13 +226,12 @@ public class MenuServiceImpl extends ServiceImpl implements MenuService {
             map.put("是否外链", item.getFrame());
             map.put("缓存", item.getCached());
             map.put("隐藏", item.getHidden());
-            map.put("权限表达式", item.getExpression());
-            map.put("是否禁用：1 表示禁用；0 表示启用", item.getDisabled());
             map.put("备注", item.getNote());
+            map.put("是否禁用：1 表示禁用；0 表示启用", item.getDisabled());
             map.put("创建者ID", item.getCreatorId());
             map.put("更新者ID", item.getUpdaterId());
             map.put("创建者名称", item.getCreatorName());
-            map.put("更新者用户名", item.getUpdaterName());
+            map.put("更新者名称", item.getUpdaterName());
             map.put("创建时间", item.getCreateTime());
             map.put("更新时间", item.getUpdateTime());
             mapList.add(map);
@@ -177,7 +246,8 @@ public class MenuServiceImpl extends ServiceImpl implements MenuService {
         }
 
         MenuDto dto = BeanUtil.toBean(entity, MenuDto.class);
-        dto.setLabel(dto.getTitle());
+        dto.setLabel(dto.getName());
+        dto.setHasChildren(dto.getChildrenNum() > 0);
         return dto;
     }
 
@@ -197,34 +267,51 @@ public class MenuServiceImpl extends ServiceImpl implements MenuService {
     }
 
     @Override
-    public List<MenuDto> treeListByIds(Collection<Serializable> ids) {
-        if (CollectionUtil.isEmpty(ids)) {
+    public List<MenuDto> treeListByPid(Long pid) {
+        if (pid == null) {
+            return Collections.emptyList();
+        }
+
+        List<MenuDto> list = pojoListByPid(pid);
+        if (CollectionUtil.isNotEmpty(list)) {
+            List<MenuDto> validList = list.stream()
+                                          .filter(Objects::nonNull)
+                                          .distinct()
+                                          .collect(Collectors.toList());
+            return buildTreeList(validList);
+        }
+        return Collections.emptyList();
+    }
+
+    private List<MenuDto> pojoListByPid(Long pid) {
+        if (pid == null) {
             return Collections.emptyList();
         }
 
         List<MenuDto> list = new ArrayList<>();
-        for (Serializable id : ids) {
-            MenuDto entity = pojoById(id);
-            if (entity == null) {
-                continue;
-            }
 
-            if (entity.getPid() != null) {
-                // 获取上级菜单
-                MenuDto parent = pojoById(entity.getPid());
-                list.add(parent);
-
-                // 获取所有同级菜单
-                MenuQuery query = new MenuQuery();
-                query.setPid(entity.getPid());
-                list.addAll(pojoListByQuery(query));
-            }
-        }
-
-        if (CollectionUtil.isEmpty(list)) {
+        // 查找父节点
+        MenuDto parentNode = pojoById(pid);
+        if (parentNode == null) {
             return Collections.emptyList();
         }
-        return buildTreeList(list);
+        parentNode.setIsDefaultExpanded(true);
+        list.add(parentNode);
+
+        MenuQuery query = new MenuQuery();
+        query.setPid(pid);
+        List<MenuDto> children = menuDao.pojoListByQuery(query, this::doToDto);
+        if (CollectionUtil.isNotEmpty(children)) {
+            list.addAll(children);
+        }
+
+        if (pid != 0) {
+            List<MenuDto> brothers = pojoListByPid(parentNode.getPid());
+            if (CollectionUtil.isNotEmpty(brothers)) {
+                list.addAll(brothers);
+            }
+        }
+        return list;
     }
 
     @Override
@@ -234,20 +321,16 @@ public class MenuServiceImpl extends ServiceImpl implements MenuService {
         }
 
         // 查找所有角色绑定的菜单
-        Set<Long> menuIds = new HashSet<>();
-        for (Long roleId : roleIds) {
-            RoleMenuMap roleMenu = new RoleMenuMap();
-            roleMenu.setRoleId(roleId);
-            List<RoleMenuMap> rolesMenus = roleMenuDao.list(Wrappers.query(roleMenu));
-            menuIds.addAll(rolesMenus.stream().map(RoleMenuMap::getMenuId).collect(Collectors.toSet()));
-        }
-
+        Set<Long> menuIds = permissionService.getMenuIdsByRoleIds(roleIds);
         if (CollectionUtil.isEmpty(menuIds)) {
             return Collections.emptyList();
         }
 
         // 根据菜单 ID 查询菜单详细信息
         List<Menu> list = menuDao.listByIds(menuIds);
+        if (CollectionUtil.isEmpty(list)) {
+            return Collections.emptyList();
+        }
 
         // 过滤菜单项
         Set<MenuDto> set = list.stream()
@@ -258,6 +341,9 @@ public class MenuServiceImpl extends ServiceImpl implements MenuService {
     }
 
     private List<MenuDto> buildTreeList(Collection<MenuDto> list) {
+        if (CollectionUtil.isEmpty(list)) {
+            return Collections.emptyList();
+        }
         TreeNodeConfig treeNodeConfig = new TreeNodeConfig();
         treeNodeConfig.setPidKey("pid");
         treeNodeConfig.setDeep(10);
@@ -265,18 +351,18 @@ public class MenuServiceImpl extends ServiceImpl implements MenuService {
         treeNodeConfig.setSort(Node.SORT.ASC);
         List<MenuDto> menuList = TreeUtil.build(list, 0L, treeNodeConfig, MenuDto.class);
         menuList.forEach(i -> {
-            i.setHasChildren(i.getSubCount() > 0);
+            i.setHasChildren(i.getChildrenNum() > 0);
             i.setLeaf(!i.isHasChildren());
         });
         return menuList;
     }
 
     @Override
-    public List<MenuVo> buildMenuListForCurrentUser() {
+    public List<VueElementMenuVo> buildMineMenuList() {
         Long userId = securityService.getCurrentUserId();
         List<RoleDto> roles = roleService.pojoListByUserId(userId);
         if (CollectionUtil.isEmpty(roles)) {
-            return new ArrayList<>();
+            return Collections.emptyList();
         }
 
         Set<Long> roleIds = roles.stream()
@@ -284,43 +370,42 @@ public class MenuServiceImpl extends ServiceImpl implements MenuService {
                                  .map(RoleDto::getId)
                                  .collect(Collectors.toSet());
         if (CollectionUtil.isEmpty(roleIds)) {
-            return new ArrayList<>();
+            return Collections.emptyList();
         }
 
         List<MenuDto> treeList = pojoTreeListByRoleIds(roleIds);
         return buildFrontMenus(treeList);
     }
 
-    private List<MenuVo> buildFrontMenus(Collection<MenuDto> list) {
-        List<MenuVo> finalVoList = new LinkedList<>();
+    private List<VueElementMenuVo> buildFrontMenus(Collection<MenuDto> list) {
+        List<VueElementMenuVo> finalVoList = new LinkedList<>();
         for (MenuDto dto : list) {
             if (dto == null) {
                 continue;
             }
 
             Collection<MenuDto> children = dto.getChildren();
-            MenuVo menuVo = new MenuVo();
-            menuVo.setName(
-                ObjectUtil.isNotEmpty(dto.getComponentName()) ? dto.getComponentName() : dto.getName());
+            VueElementMenuVo menuVo = new VueElementMenuVo();
+            menuVo.setName(ObjectUtil.isNotEmpty(dto.getComponent()) ? dto.getComponent() : dto.getName());
             // 一级目录需要加斜杠，不然会报警告
-            menuVo.setPath(dto.getPid() == null ? "/" + dto.getPath() : dto.getPath());
+            menuVo.setPath(dto.getPid() == 0 ? "/" + dto.getPath() : dto.getPath());
             menuVo.setHidden(dto.getHidden());
             // 如果不是外链
             if (!dto.getFrame()) {
-                if (dto.getPid() == null) {
+                if (dto.getPid() == 0) {
                     menuVo.setComponent(StrUtil.isEmpty(dto.getComponent()) ? "Layout" : dto.getComponent());
                 } else if (!StrUtil.isEmpty(dto.getComponent())) {
                     menuVo.setComponent(dto.getComponent());
                 }
             }
-            menuVo.setMeta(new MenuVo.MenuMetaVo(dto.getTitle(), dto.getIcon(), !dto.getCached()));
+            menuVo.setMeta(new VueElementMenuVo.MenuMetaVo(dto.getName(), dto.getIcon(), !dto.getCached()));
             if (CollectionUtil.isNotEmpty(children)) {
                 menuVo.setAlwaysShow(true);
                 menuVo.setRedirect("noredirect");
                 menuVo.setChildren(buildFrontMenus(children));
                 // 处理是一级菜单并且没有子菜单的情况
-            } else if (dto.getPid() == null) {
-                MenuVo menuVo1 = new MenuVo();
+            } else if (dto.getPid() == 0) {
+                VueElementMenuVo menuVo1 = new VueElementMenuVo();
                 menuVo1.setMeta(menuVo.getMeta());
                 // 非外链
                 if (!dto.getFrame()) {
@@ -333,7 +418,7 @@ public class MenuServiceImpl extends ServiceImpl implements MenuService {
                 menuVo.setName(null);
                 menuVo.setMeta(null);
                 menuVo.setComponent("Layout");
-                List<MenuVo> list1 = new ArrayList<>();
+                List<VueElementMenuVo> list1 = new ArrayList<>();
                 list1.add(menuVo1);
                 menuVo.setChildren(list1);
             }
@@ -365,24 +450,6 @@ public class MenuServiceImpl extends ServiceImpl implements MenuService {
         }
 
         return ids;
-    }
-
-    private Object treeObject(List<Menu> menus) {
-        List<Map<String, Object>> list = new LinkedList<>();
-        menus.forEach(menu -> {
-                if (menu != null) {
-                    List<Menu> menuList = menuDao.listByPid(menu.getId());
-                    Map<String, Object> map = new HashMap<>(16);
-                    map.put("id", menu.getId());
-                    map.put("label", menu.getName());
-                    if (menuList != null && menuList.size() != 0) {
-                        map.put("children", treeObject(menuList));
-                    }
-                    list.add(map);
-                }
-            }
-        );
-        return list;
     }
 
 }
